requirements.txt
fastapi
uvicorn
pydantic
requests
qdrant-client
sentence-transformers
pydantic-settings
starlette
asyncpg
httpx
